requirements.txt
# LLM & RAG
transformers>=4.40
sentence-transformers>=2.7
chromadb>=0.5
faiss-cpu>=1.8
bitsandbytes>=0.43        # 4-bit/8-bit quantization
accelerate>=0.29          # Required for device_map and quantization

# PDF Parsing
pymupdf>=1.24
pdfplumber>=0.11
pytesseract>=0.3          # OCR for scanned PDFs
Pillow>=10.0              # Required by pytesseract

# API server
fastapi>=0.111
uvicorn>=0.29
python-multipart>=0.0.9   # PDF file upload support
sse-starlette>=2.0        # Server-Sent Events for chat streaming
httpx>=0.27               # Async HTTP client

# Evaluation
ragas>=0.1
datasets>=2.19
openai>=1.0               # Required by ragas for LLM judge

# Utilities
pyyaml>=6.0
python-dotenv>=1.0
torch>=2.1                # PyTorch for model inference
